Move css styles to seperated dir
