+++ src/styles/AvatarPicker.css
